Extracted tag group module
