+++ src/app/component/admin/tag-group/delete/delete.component.spec.ts
@@ -3,7 +3,6 @@ import { ActivatedRoute, Router } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { ApiErrorDetails } from "@baw-api/api.interceptor.service";
 import { tagGroupResolvers, TagGroupService } from "@baw-api/tag-group.service";
-import { adminTagGroupsMenuItem } from "@component/admin/admin.menus";
 import { TagGroup } from "@models/TagGroup";
 import { SharedModule } from "@shared/shared.module";
 import { ToastrService } from "ngx-toastr";
@@ -11,6 +10,7 @@ import { BehaviorSubject } from "rxjs";
 import { appLibraryImports } from "src/app/app.module";
 import { mockActivatedRoute, testBawServices } from "src/app/test.helper";
 import { assertFormErrorHandling } from "src/testHelpers";
+import { adminTagGroupsMenuItem } from "../tag-group.menus";
 import { AdminTagGroupsDeleteComponent } from "./delete.component";
 
 describe("AdminTagGroupsDeleteComponent", () => {
