+++ src/app/component/admin/tag-group/delete/delete.component.ts
@@ -2,11 +2,6 @@ import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 import { tagGroupResolvers, TagGroupService } from "@baw-api/tag-group.service";
 import {
-  adminDeleteTagGroupMenuItem,
-  adminTagGroupsCategory,
-  adminTagGroupsMenuItem,
-} from "@component/admin/admin.menus";
-import {
   defaultSuccessMsg,
   FormTemplate,
 } from "@helpers/formTemplate/formTemplate";
@@ -15,6 +10,11 @@ import { TagGroup } from "@models/TagGroup";
 import { List } from "immutable";
 import { ToastrService } from "ngx-toastr";
 import { adminTagGroupMenuItemActions } from "../list/list.component";
+import {
+  adminDeleteTagGroupMenuItem,
+  adminTagGroupsCategory,
+  adminTagGroupsMenuItem,
+} from "../tag-group.menus";
 
 const tagGroupKey = "tagGroup";
