+++ src/app/component/admin/tag-group/edit/edit.component.ts
@@ -2,11 +2,6 @@ import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 import { tagGroupResolvers, TagGroupService } from "@baw-api/tag-group.service";
 import {
-  adminEditTagGroupMenuItem,
-  adminTagGroupsCategory,
-  adminTagGroupsMenuItem,
-} from "@component/admin/admin.menus";
-import {
   defaultSuccessMsg,
   FormTemplate,
 } from "@helpers/formTemplate/formTemplate";
@@ -16,6 +11,11 @@ import { List } from "immutable";
 import { ToastrService } from "ngx-toastr";
 import { adminTagGroupMenuItemActions } from "../list/list.component";
 import { fields } from "../tag-group.json";
+import {
+  adminEditTagGroupMenuItem,
+  adminTagGroupsCategory,
+  adminTagGroupsMenuItem,
+} from "../tag-group.menus";
 
 const tagGroupKey = "tagGroup";
